refactor: move service accessors
