+++ docs/import_dependency_map.md
@@ -20,8 +20,8 @@ app.py
 Key points:
 
 * Configuration no longer resolves optional services at import time.  Functions
-  such as `get_database_manager()` retrieve services from the registry when
-  called, preventing early imports.
+  such as `config.service_integration.get_database_manager()` retrieve services
+  from the registry when called, preventing early imports.
 * `security_callback_controller` imports `CallbackEvent` from the lightweight
   `core.callback_events` module.  This removes a link in the chain that used to
   pull in the entire callback controller during start-up.
